+++ tests/unit/FillerStack_test.go
@@ -15,25 +15,21 @@ func TestFillerStack(t *testing.T) {
 
 	grammar := internalGrammarParser.CreateGrammar(`[
 
-		{ rule: s(P1) -> which() np(E1) dep_vp(P1, E1),			sense: which(E1) }
+		{ rule: s(P1) -> 'which' np(E1) dep_vp(P1, E1),			sense: which(E1) }
 		{ rule: np(E1) -> nbar(E1), 							sense: quant(_, some(_), E1, sem(1), sem(parent)) }
 		{ rule: nbar(E) -> noun(E) }
-		{ rule: dep_vp(P1, E1) -> be() np(E2) advp(P1) vp(P1, E1, E2) }
+		{ rule: dep_vp(P1, E1) -> be(_) np(E2) advp(P1) vp(P1, E1, E2) }
 		{ rule: np(E1) -> qp(Q1) nbar(E1), 						sense: quant(Q1, sem(1), E1, sem(2), sem(parent)) }
 		{ rule: advp(P1) -> adverb(P1) }
-		{ rule: vp(P1, E1, E2) -> to() take() from(), 			sense: take_from(P1, E2, E1)  }
+		{ rule: vp(P1, E1, E2) -> 'to' 'take' 'from', 			sense: take_from(P1, E2, E1)  }
+		{ rule: be(P1) -> 'were' }
+		{ rule: qp(Q1) -> 'the', sense: the(Q1) }
+		{ rule: adverb(A1) -> 'easiest', sense: easiest(A1) }
+		{ rule: noun(E1) -> 'babies', sense: baby(E1) }
+		{ rule: noun(E1) -> 'toys', sense: toy(E1) }
 	]`)
 
 	lexicon := internalGrammarParser.CreateLexicon(`[
-		{ form: 'which', pos: which }
-		{ form: 'were', pos: be }
-		{ form: 'babies', pos: noun, 		sense: baby(E) }
-		{ form: 'toys', pos: noun, 			sense: toy(E) }
-		{ form: 'the', pos: qp, 				sense: the(E) }
-		{ form: 'easiest', pos: adverb, 		sense: easiest(E) }
-		{ form: 'to', pos: to }
-		{ form: 'take', pos: take }
-		{ form: 'from', pos: from }
 	]`)
 	log := common.NewSystemLog(true)